Bot v2
made major changes to the original source code to be easier to read, more terse and scalable with the help of a colleague (thanks Ken).1) removed the individual function for each player and simplified the functions to be reused in the cocde. 2) reworked the function to calculate win/lose streaks (this is much better now). 3) made some minor changes to how data is displayed in the embedded discord message